Add Categorie Api

+++ src/main/java/com/restaurantUniversitaire/Cantine/El/Ons/Persistance/dao/PaiementRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PaiementRepository extends JpaRepository<Paiement, Long> {
+public interface PaiementRepository extends JpaRepository<Paiement, Integer> {
 }
